test changing target butcast

+++ tinygrad/renderer/cstyle.py
@@ -379,7 +379,7 @@ code_for_op_hip = { UnaryOps.SQRT: lambda x,dtype: f"__ocml_sqrt_f{ {dtypes.half
   # }
 
 def cast_float_bf16(x: UOp) -> UOp:
-  u32 = x.bitcast(dtypes.uint32)
+  u32 = x.bitcast(dtypes.float)
 
   is_not_inf_nan = (-u32) & 0x7f800000
   has_mantissa = u32 & 0xffff
@@ -413,8 +413,6 @@ class AMDRenderer(CStyleLanguage):
   float4 = "make_float4"
   uses_ptr_arithmetic = False  # NOTE: this fixes TestLinearizerOverflowAlt
   type_map = {dtypes.bfloat16: "hip_bfloat16"}
-  # string_rewrite = PatternMatcher([
-  #   (UPat(UOps.BITCAST, name="x"), lambda r,x: f"*reinterpret_cast<{r.render_dtype(x.dtype)}*>(&{r[x.src[0]]})")]) + base_rewrite
 
   extra_matcher = PatternMatcher([
     (UPat(UOps.ALU, arg=TernaryOps.WHERE, src=(UPat.var("b"), UPat.var("x", dtype=dtypes.bfloat16), UPat.var("y", dtype=dtypes.bfloat16))),
